Move to using fewer files across the codebase
This is no code changes, purely moving code around
