--- client/environment_variables_delete.go
@@ -1,25 +1,0 @@
-package client
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/hashicorp/terraform-plugin-log/tflog"
-)
-
-// DeleteEnvironmentVariable will remove an environment variable from Vercel.
-func (c *Client) DeleteEnvironmentVariable(ctx context.Context, projectID, teamID, variableID string) error {
-	url := fmt.Sprintf("%s/v8/projects/%s/env/%s", c.baseURL, projectID, variableID)
-	if c.teamID(teamID) != "" {
-		url = fmt.Sprintf("%s?teamId=%s", url, c.teamID(teamID))
-	}
-	tflog.Info(ctx, "deleting environment variable", map[string]interface{}{
-		"url": url,
-	})
-	return c.doRequest(clientRequest{
-		ctx:    ctx,
-		method: "DELETE",
-		url:    url,
-		body:   "",
-	}, nil)
-}
