+++ client/team.go
@@ -35,3 +35,32 @@ func (c *Client) CreateTeam(ctx context.Context, request TeamCreateRequest) (r T
 	}, &r)
 	return r, err
 }
+
+// DeleteTeam deletes an existing team within vercel.
+func (c *Client) DeleteTeam(ctx context.Context, teamID string) error {
+	url := fmt.Sprintf("%s/v1/teams/%s", c.baseURL, teamID)
+	tflog.Info(ctx, "deleting team", map[string]interface{}{
+		"url": url,
+	})
+	return c.doRequest(clientRequest{
+		ctx:    ctx,
+		method: "DELETE",
+		url:    url,
+		body:   "",
+	}, nil)
+}
+
+// GetTeam returns information about an existing team within vercel.
+func (c *Client) GetTeam(ctx context.Context, idOrSlug string) (r TeamResponse, err error) {
+	url := fmt.Sprintf("%s/v2/teams/%s", c.baseURL, idOrSlug)
+	tflog.Info(ctx, "getting team", map[string]interface{}{
+		"url": url,
+	})
+	err = c.doRequest(clientRequest{
+		ctx:    ctx,
+		method: "GET",
+		url:    url,
+		body:   "",
+	}, &r)
+	return r, err
+}
